+++ R/class_pipeline.R
@@ -1,10 +1,8 @@
-pipeline_init <- function(targets = list()) {
-  targets <- pipeline_targets_init(targets)
-  envir <- pipeline_envir(targets)
-  imports <- imports_init(envir)
+pipeline_init <- function(targets = list(), clone_targets = TRUE) {
+  targets <- pipeline_targets_init(targets, clone_targets)
+  imports <- imports_init(tar_option_get("envir"))
   pipeline_new(
     targets = targets,
-    envir = envir,
     imports = imports,
     loaded = counter_init(),
     transient = counter_init()
@@ -13,37 +11,30 @@ pipeline_init <- function(targets = list()) {
 
 pipeline_new <- function(
   targets = NULL,
-  envir = NULL,
   imports = NULL,
   loaded = NULL,
   transient = NULL
 ) {
   force(targets)
-  force(envir)
   force(imports)
   force(loaded)
   force(transient)
   enclass(environment(), "tar_pipeline")
 }
 
-pipeline_targets_init <- function(targets) {
-  targets <- targets %||% list()
+pipeline_targets_init <- function(targets, clone_targets) {
+  targets <- targets %|||% list()
   assert_target_list(targets)
   names <- map_chr(targets, ~.x$settings$name)
   assert_unique_targets(names)
+  if (clone_targets) {
+    # If the user has target objects in the global environment,
+    # loading data into them may cause huge data transfers to workers.
+    # Best to not modify the user's copies of target objects.
+    targets <- map(targets, ~target_subpipeline_copy(.x, keep_value = FALSE))
+  }
   names(targets) <- names
   list2env(targets, parent = emptyenv(), hash = TRUE)
-}
-
-pipeline_envir <- function(targets) {
-  names <- names(targets)
-  out <- trn(
-    length(names),
-    targets[[names[1]]]$cache$imports$envir,
-    tar_empty_envir
-  )
-  assert_envir(out, "pipeline must initialize with only stems and patterns.")
-  out
 }
 
 pipeline_get_target <- function(pipeline, name) {
@@ -92,7 +83,7 @@ pipeline_set_target <- function(pipeline, target) {
 }
 
 pipeline_exists_target <- function(pipeline, name) {
-  envir <- pipeline$targets %||% tar_empty_envir
+  envir <- pipeline$targets %|||% tar_empty_envir
   exists(x = name, envir = envir, inherits = FALSE)
 }
 
@@ -113,7 +104,7 @@ pipeline_upstream_edges <- function(pipeline, targets_only = TRUE) {
   edge_list <- map(pipeline$targets, ~target_upstream_edges(.x))
   edges <- do.call(rbind, edge_list)
   edges <- trn(targets_only, pipeline_targets_only_edges(edges), edges)
-  edges <- edges %||% data_frame(from = character(0), to = character(0))
+  edges <- edges %|||% data_frame(from = character(0), to = character(0))
   rownames(edges) <- NULL
   edges
 }
@@ -171,7 +162,7 @@ pipeline_produce_subpipeline <- function(pipeline, name, keep_value = NULL) {
   target <- pipeline_get_target(pipeline, name)
   deps <- target_deps_deep(target, pipeline)
   targets <- new.env(parent = emptyenv())
-  keep_value <- keep_value %||% identical(target$settings$retrieval, "main")
+  keep_value <- keep_value %|||% identical(target$settings$retrieval, "main")
   lapply(
     deps,
     pipeline_assign_target_copy,
@@ -181,7 +172,6 @@ pipeline_produce_subpipeline <- function(pipeline, name, keep_value = NULL) {
   )
   pipeline_new(
     targets = targets,
-    envir = tar_empty_envir,
     loaded = counter_init(),
     transient = counter_init()
   )
@@ -238,19 +228,6 @@ pipeline_validate_dag <- function(igraph) {
   }
 }
 
-pipeline_validate_envirs <- function(pipeline) {
-  assert_envir(pipeline$envir)
-  assert_envir(pipeline$imports %||% tar_empty_envir)
-  targets <- as.list(pipeline$targets)
-  lapply(targets, pipeline_validate_envir, envir = pipeline$envir)
-}
-
-pipeline_validate_envir <- function(target, envir) {
-  if (!identical(target$cache$imports$envir, envir)) {
-    throw_validate("all targets must share the same environment")
-  }
-}
-
 pipeline_validate_conflicts <- function(pipeline) {
   conflicts <- intersect(names(pipeline$imports), names(pipeline$targets))
   msg <- paste0(
@@ -281,7 +258,6 @@ pipeline_validate <- function(pipeline) {
 pipeline_validate_lite <- function(pipeline) {
   assert_inherits(pipeline, "tar_pipeline", msg = "invalid pipeline.")
   assert_correct_fields(pipeline, pipeline_new)
-  pipeline_validate_envirs(pipeline)
   pipeline_validate_conflicts(pipeline)
 }