+++ R/tar_manifest.R
@@ -61,7 +61,7 @@ tar_manifest <- function(
   names = NULL,
   fields = c("name", "command", "pattern"),
   callr_function = callr::r,
-  callr_arguments = list()
+  callr_arguments = targets::callr_args_default(callr_function)
 ) {
   assert_script()
   assert_callr_function(callr_function)
@@ -85,11 +85,11 @@ tar_manifest_inner <- function(
 ) {
   igraph <- pipeline_produce_igraph(pipeline, targets_only = TRUE)
   all_names <- topo_sort_igraph(igraph)
-  names <- eval_tidyselect(names_quosure, all_names) %||% all_names
+  names <- eval_tidyselect(names_quosure, all_names) %|||% all_names
   names <- intersect(all_names, names)
   out <- map(names, ~tar_manifest_target(pipeline_get_target(pipeline, .x)))
   out <- do.call(rbind, out)
-  fields <- eval_tidyselect(fields_quosure, colnames(out)) %||% colnames(out)
+  fields <- eval_tidyselect(fields_quosure, colnames(out)) %|||% colnames(out)
   out[, base::union("name", fields), drop = FALSE]
 }