+++ R/tar_option_set.R
@@ -19,7 +19,42 @@
 #'   Similarly, objects in `tar_option_get("envir")` override
 #'   everything in `tar_option_get("imports")`.
 #' @param envir Environment containing functions and global objects
-#'   used in the R commands to run targets.
+#'   used in the R commands to run targets. Defaults to the global environment.
+#'   If `envir` is the global environment, all the promise objects
+#'   are diffused before sending the data to parallel workers
+#'   in [tar_make_future()] and [tar_make_clustermq()],
+#'   but otherwise the environment is unmodified.
+#'   This behavior improves performance by decreasing
+#'   the size of data sent to workers.
+#'
+#'   If `envir` is not the global environment, then it should at least inherit
+#'   from the global environment or base environment
+#'   so `targets` can access attached packages.
+#'   In the case of a non-global `envir`, `targets` attempts to remove
+#'   potentially high memory objects that come directly from `targets`.
+#'   That includes `tar_target()` objects of class `"tar_target"`,
+#'   as well as objects of class `"tar_pipeline"` or `"tar_algorithm"`.
+#'   This behavior improves performance by decreasing
+#'   the size of data sent to workers.
+#'
+#'   Package environments should not be assigned to `envir`.
+#'   To include package objects as upstream dependencies in the pipeline,
+#'   assign the package to the `packages` and `imports` arguments
+#'   of `tar_option_set()`.
+#' @param backoff Numeric of length 1, must be greater than or equal to 0.01.
+#'   Maximum upper bound of the random polling interval
+#'   for the priority queue (seconds).
+#'   In high-performance computing (e.g. [tar_make_clustermq()]
+#'   and [tar_make_future()]) it can be expensive to repeatedly poll the
+#'   priority queue if no targets are ready to process. The number of seconds
+#'   between polls is `runif(1, 0.01, max(backoff, 0.01 * 1.5 ^ index))`,
+#'   where `index` is the number of consecutive polls so far that found
+#'   no targets ready to skip or run.
+#'   (If no target is ready, `index` goes up by 1. If a target is ready,
+#'   `index` resets to 0. For more information on exponential,
+#'   backoff, visit <https://en.wikipedia.org/wiki/Exponential_backoff>).
+#'   Raising `backoff` is kinder to the CPU etc. but may incur delays
+#'   in some instances.
 #' @param debug Character vector of names of targets to run in debug mode.
 #'   To use effectively, you must set `callr_function = NULL` and
 #'   restart your R session just before running. You should also
@@ -67,6 +102,7 @@ tar_option_set <- function(
   garbage_collection = NULL,
   deployment = NULL,
   priority = NULL,
+  backoff = NULL,
   resources = NULL,
   storage = NULL,
   retrieval = NULL,
@@ -87,6 +123,7 @@ tar_option_set <- function(
   tar_option_set_garbage_collection(garbage_collection)
   tar_option_set_deployment(deployment)
   tar_option_set_priority(priority)
+  tar_option_set_backoff(backoff)
   tar_option_set_resources(resources)
   tar_option_set_storage(storage)
   tar_option_set_retrieval(retrieval)
@@ -96,31 +133,31 @@ tar_option_set <- function(
 }
 
 tar_option_set_tidy_eval <- function(tidy_eval) {
-  tidy_eval <- tidy_eval %||% tar_option_get("tidy_eval")
+  tidy_eval <- tidy_eval %|||% tar_option_get("tidy_eval")
   assert_lgl(tidy_eval, "tidy_eval in tar_option_set() must be logical.")
   assign("tidy_eval", tidy_eval, envir = tar_envir_options)
 }
 
 tar_option_set_packages <- function(packages) {
-  packages <- packages %||% tar_option_get("packages")
+  packages <- packages %|||% tar_option_get("packages")
   assert_chr(packages, "packages in tar_option_set() must be character.")
   assign("packages", packages, envir = tar_envir_options)
 }
 
 tar_option_set_imports <- function(imports) {
-  imports <- imports %||% tar_option_get("imports")
+  imports <- imports %|||% tar_option_get("imports")
   assert_chr(imports, "imports in tar_option_set() must be character.")
   assign("imports", imports, envir = tar_envir_options)
 }
 
 tar_option_set_library <- function(library) {
-  library <- library %||% tar_option_get("library")
-  assert_chr(library %||% character(0), "library must be NULL or character.")
+  library <- library %|||% tar_option_get("library")
+  assert_chr(library %|||% character(0), "library must be NULL or character.")
   assign("library", library, envir = tar_envir_options)
 }
 
 tar_option_set_envir <- function(envir) {
-  envir <- envir %||% tar_option_get("envir")
+  envir <- envir %|||% tar_option_get("envir")
   msg <- paste(
     "envir in tar_option_set() must be the environment",
     "where you put your functions and global objects",
@@ -131,31 +168,31 @@ tar_option_set_envir <- function(envir) {
 }
 
 tar_option_set_format <- function(format) {
-  format <- format %||% tar_option_get("format")
+  format <- format %|||% tar_option_get("format")
   assert_format(format)
   assign("format", format, envir = tar_envir_options)
 }
 
 tar_option_set_iteration <- function(iteration) {
-  iteration <- iteration %||% tar_option_get("iteration")
+  iteration <- iteration %|||% tar_option_get("iteration")
   assert_flag(iteration, c("vector", "list", "group"))
   assign("iteration", iteration, envir = tar_envir_options)
 }
 
 tar_option_set_error <- function(error) {
-  error <- error %||% tar_option_get("error")
+  error <- error %|||% tar_option_get("error")
   assert_flag(error, c("stop", "continue", "workspace"))
   assign("error", error, envir = tar_envir_options)
 }
 
 tar_option_set_memory <- function(memory) {
-  memory <- memory %||% tar_option_get("memory")
+  memory <- memory %|||% tar_option_get("memory")
   assert_flag(memory, c("persistent", "transient"))
   assign("memory", memory, envir = tar_envir_options)
 }
 
 tar_option_set_garbage_collection <- function(garbage_collection) {
-  garbage_collection <- garbage_collection %||%
+  garbage_collection <- garbage_collection %|||%
     tar_option_get("garbage_collection")
   garbage_collection <- as.logical(garbage_collection)
   assert_lgl(garbage_collection, "garbage_collection must be logical.")
@@ -164,13 +201,13 @@ tar_option_set_garbage_collection <- function(garbage_collection) {
 }
 
 tar_option_set_deployment <- function(deployment) {
-  deployment <- deployment %||% tar_option_get("deployment")
+  deployment <- deployment %|||% tar_option_get("deployment")
   assert_flag(deployment, c("worker", "main"))
   assign("deployment", deployment, envir = tar_envir_options)
 }
 
 tar_option_set_priority <- function(priority) {
-  priority <- priority %||% tar_option_get("priority")
+  priority <- priority %|||% tar_option_get("priority")
   assert_dbl(priority, msg = "priority must be numeric")
   assert_scalar(priority, msg = "priority must have length 1")
   assert_ge(priority, 0, msg = "priority cannot be less than 0")
@@ -178,38 +215,47 @@ tar_option_set_priority <- function(priority) {
   assign("priority", priority, envir = tar_envir_options)
 }
 
+tar_option_set_backoff <- function(backoff) {
+  backoff <- backoff %|||% tar_option_get("backoff")
+  assert_dbl(backoff, msg = "backoff must be numeric")
+  assert_scalar(backoff, msg = "backoff must have length 1")
+  assert_ge(backoff, 0.01, msg = "backoff cannot be less than 0.01")
+  assert_le(backoff, 1e9, msg = "backoff cannot be greater than 1e9")
+  assign("backoff", backoff, envir = tar_envir_options)
+}
+
 tar_option_set_resources <- function(resources) {
-  resources <- resources %||% tar_option_get("resources")
+  resources <- resources %|||% tar_option_get("resources")
   assert_list(resources, "resources in tar_option_set() must be a named list.")
   assign("resources", resources, envir = tar_envir_options)
 }
 
 tar_option_set_storage <- function(storage) {
-  storage <- storage %||% tar_option_get("storage")
+  storage <- storage %|||% tar_option_get("storage")
   assert_flag(storage, c("main", "worker"))
   assign("storage", storage, envir = tar_envir_options)
 }
 
 tar_option_set_retrieval <- function(retrieval) {
-  retrieval <- retrieval %||% tar_option_get("retrieval")
+  retrieval <- retrieval %|||% tar_option_get("retrieval")
   assert_flag(retrieval, c("main", "worker"))
   assign("retrieval", retrieval, envir = tar_envir_options)
 }
 
 tar_option_set_cue <- function(cue) {
-  cue <- cue %||% tar_option_get("cue")
+  cue <- cue %|||% tar_option_get("cue")
   trn(is.null(cue), NULL, cue_validate(cue))
   assign("cue", cue, envir = tar_envir_options)
 }
 
 tar_option_set_debug <- function(debug) {
-  debug <- debug %||% tar_option_get("debug")
+  debug <- debug %|||% tar_option_get("debug")
   assert_chr(debug, "debug argument of tar_option_set() must be a character.")
   assign("debug", debug, envir = tar_envir_options)
 }
 
 tar_option_set_workspaces <- function(workspaces) {
-  workspaces <- workspaces %||% tar_option_get("workspaces")
+  workspaces <- workspaces %|||% tar_option_get("workspaces")
   assert_chr(
     workspaces,
     "workspaces argument of tar_option_set() must be a character."