+++ R/tar_watch.R
@@ -43,7 +43,7 @@
 #' })
 #' }
 tar_watch <- function(
-  seconds = 5,
+  seconds = 15,
   seconds_min = 1,
   seconds_max = 60,
   seconds_step = 1,
@@ -68,8 +68,8 @@ tar_watch <- function(
     "shinyWidgets",
     "visNetwork"
   )
-  msg <- paste("tar_watch() requires packages", paste(pkgs, collapse = ", "))
-  map(pkgs, ~assert_package(.x, msg = msg))
+
+  assert_package(pkgs)
   assert_dbl(seconds, "seconds must be numeric.")
   assert_dbl(seconds_min, "seconds_min must be numeric.")
   assert_dbl(seconds_max, "seconds_max must be numeric.")
@@ -250,9 +250,20 @@ tar_watch_ui <- function(
         choiceValues = c("graph", "branches"),
         selected = "graph"
       ),
-      shinyWidgets::materialSwitch(
+      shinyWidgets::actionBttn(
         inputId = ns("refresh"),
         label = "refresh",
+        style = "simple",
+        color = "primary",
+        size = "sm",
+        block = FALSE,
+        no_outline = TRUE
+      ),
+      shiny::br(),
+      shiny::br(),
+      shinyWidgets::materialSwitch(
+        inputId = ns("watch"),
+        label = "watch",
         value = TRUE,
         status = "primary",
         right = TRUE
@@ -330,25 +341,27 @@ tar_watch_server <- function(id, height = "650px") {
   shiny::moduleServer(
     id,
     function(input, output, session) {
-      interval <- 1000
-      react_refresh <- shiny::reactive(input$refresh)
-      react_display <- shiny::reactive(input$display)
+      interval <- 200
+      refresh <- shiny::reactiveValues(refresh = tempfile())
       react_millis <- shiny::reactive(1000 * as.numeric(input$seconds))
       react_targets <- shiny::reactive(as.logical(input$targets_only))
       react_outdated <- shiny::reactive(as.logical(input$outdated))
       react_label <- shiny::reactive(input$label)
       react_ls <- shiny::reactive(as.numeric(input$level_separation))
-      display <- shiny::throttle(r = react_display, millis = interval)
-      refresh <- shiny::throttle(r = react_refresh, millis = interval)
-      millis <- shiny::throttle(r = react_millis, millis = interval)
-      targets_only <- shiny::throttle(r = react_targets, millis = interval)
-      outdated_tl <- shiny::throttle(r = react_outdated, millis = interval)
-      label <- shiny::throttle(r = react_label, millis = interval)
-      level_separation <- shiny::throttle(r = react_ls, millis = interval)
-      output$graph <- visNetwork::renderVisNetwork({
-        if (identical(react_refresh(), TRUE)) {
+      millis <- shiny::debounce(r = react_millis, millis = interval)
+      targets_only <- shiny::debounce(r = react_targets, millis = interval)
+      outdated_tl <- shiny::debounce(r = react_outdated, millis = interval)
+      label <- shiny::debounce(r = react_label, millis = interval)
+      level_separation <- shiny::debounce(r = react_ls, millis = interval)
+      shiny::observe({
+        if (identical(input$watch, TRUE)) {
           shiny::invalidateLater(millis = millis())
+          refresh$refresh <- tempfile()
         }
+      })
+      shiny::observeEvent(input$refresh, refresh$refresh <- tempfile())
+      output$graph <- visNetwork::renderVisNetwork({
+        shiny::req(refresh$refresh)
         trn(
           tar_exist_script(),
           tar_visnetwork(
@@ -367,9 +380,7 @@ tar_watch_server <- function(id, height = "650px") {
         )
       })
       output$branches <- gt::render_gt({
-        if (identical(react_refresh(), TRUE)) {
-          shiny::invalidateLater(millis = millis())
-        }
+        shiny::req(refresh$refresh)
         trn(
           tar_exist_progress(),
           tar_progress_branches_gt(),
@@ -378,7 +389,7 @@ tar_watch_server <- function(id, height = "650px") {
       }, height = height)
       output$display <- shiny::renderUI({
         switch(
-          display() %||% "graph",
+          input$display %|||% "graph",
           graph = shinycssloaders::withSpinner(
             visNetwork::visNetworkOutput(session$ns("graph"), height = height)
           ),