+++ tests/hpc/test-future_sge.R
@@ -1,45 +1,3 @@
-tar_test("future workers actually launch", {
-  skip_on_cran()
-  tar_script({
-    future::plan(future.callr::callr)
-    list(
-      tar_target(x, seq_len(4)),
-      tar_target(
-        y,
-        Sys.sleep(30),
-        pattern = map(x)
-      )
-    )
-  })
-  # The following should run 4 targets concurrently.
-  # Terminate early if necessary.
-  tar_make_future(workers = 4)
-  tar_progress()
-})
-
-tar_test("custom future plans through resources", {
-  skip_on_cran()
-  tar_script({
-    future::plan(future::multisession, workers = 4)
-    plan_multisession <- future::plan()
-    future::plan(future::sequential)
-    list(
-      tar_target(x, seq_len(4)),
-      tar_target(
-        y,
-        Sys.sleep(30),
-        pattern = map(x),
-        resources = list(plan = plan_multisession)
-      )
-    )
-  })
-  # The following should run 4 targets concurrently.
-  tar_make_future(workers = 4)
-  # After all 4 targets start, terminate the pipeline early and show progress.
-  # x should be built, and y and its 4 branches should be listed as started.
-  tar_progress()
-})
-
 test_that("packages are actually loaded", {
   # Needs sge_batchtools.tmpl (in current directory).
   unlink("_targets", recursive = TRUE)
@@ -68,34 +26,60 @@ test_that("packages are actually loaded", {
   expect_equal(tar_read(x), exp)
 })
 
-test_that("nontrivial globals", {
-  unlink("_targets", recursive = TRUE)
-  on.exit(unlink("_targets", recursive = TRUE))
+test_that("nontrivial globals with custom environment", {
+  skip_on_cran()
   skip_if_not_installed("future")
   skip_if_not_installed("future.batchtools")
-  on.exit(future::plan(future::sequential), add = TRUE)
-  future::plan(
-    future.batchtools::batchtools_sge,
-    template = "sge_batchtools.tmpl"
-  )
-  old_envir <- tar_option_get("envir")
-  on.exit(tar_option_set(envir = old_envir), add = TRUE)
-  envir <- new.env(parent = globalenv())
-  tar_option_set(envir = envir)
-  evalq({
+  tar_destroy()
+  on.exit(tar_destroy())
+  tar_script({
+    future::plan(
+      future.batchtools::batchtools_sge,
+      template = "sge_batchtools.tmpl"
+    )
+    envir <- new.env(parent = globalenv())
+    evalq({
+      f <- function(x) {
+        g(x) + 1L
+      }
+      g <- function(x) {
+        x + 1L
+      }
+    }, envir = envir)
+    tar_option_set(envir = envir)
+    list(
+      tar_target(x, 1),
+      tar_target(y, f(x))
+    )
+  })
+  tar_make_future()
+  expect_equal(tar_read(y), 3L)
+})
+
+test_that("nontrivial globals with global environment", {
+  skip_on_cran()
+  skip_if_not_installed("future")
+  skip_if_not_installed("future.batchtools")
+  tar_destroy()
+  on.exit(tar_destroy())
+  tar_script({
+    future::plan(
+      future.batchtools::batchtools_sge,
+      template = "sge_batchtools.tmpl"
+    )
     f <- function(x) {
       g(x) + 1L
     }
     g <- function(x) {
       x + 1L
     }
-  }, envir = envir)
-  x <- tar_target_raw("x", quote(f(1L)))
-  pipeline <- pipeline_init(list(x))
-  algo <- future_init(pipeline)
-  algo$run()
-  target <- pipeline_get_target(pipeline, "x")
-  expect_equal(tar_read(x), 3L)
+    list(
+      tar_target(x, 1),
+      tar_target(y, f(x))
+    )
+  })
+  tar_make_future()
+  expect_equal(tar_read(y), 3L)
 })
 
 test_that("branching plan on SGE", {