+++ tests/performance/test-memory.R
@@ -1,6 +1,67 @@
-library(targets)
-library(pryr)
+# local_init(pipeline)$run() retains correct objects in persistent memory
+library(testthat)
+tar_destroy()
+tar_option_set(envir = environment(), memory = "persistent")
+pipeline <- pipeline_init(
+  list(
+    tar_target_raw(name = "data1", quote(seq_len(10))),
+    tar_target_raw(name = "data2", quote(seq_len(20))),
+    tar_target_raw(name = "min1", quote(min(data1))),
+    tar_target_raw(name = "min2", quote(min(data2))),
+    tar_target_raw(name = "max1", quote(max(data1))),
+    tar_target_raw(name = "max2", quote(max(data2))),
+    tar_target_raw(name = "mins", quote(c(min1, min2))),
+    tar_target_raw(name = "maxes", quote(c(max1, max2))),
+    tar_target_raw(
+      name = "all",
+      quote(c(browser(), mins, maxes))
+    )
+  )
+)
+# Enter the debugger:
+local_init(pipeline)$run()
+# Run these tests inside the debugger:
+names <- pipeline_get_names(pipeline)
+result <- map_lgl(names, ~is.null(pipeline_get_target(pipeline, .x)$value))
+expect_true(result["all"])
+expect_false(any(result[setdiff(names(result), "all")]))
+# Exit the debugger.
 
+# Same test with transient memory:
+tar_destroy()
+tar_option_set(envir = environment(), memory = "transient")
+pipeline <- pipeline_init(
+  list(
+    tar_target_raw(name = "data1", quote(seq_len(10))),
+    tar_target_raw(name = "data2", quote(seq_len(20))),
+    tar_target_raw(name = "min1", quote(min(data1))),
+    tar_target_raw(name = "min2", quote(min(data2))),
+    tar_target_raw(name = "max1", quote(max(data1))),
+    tar_target_raw(name = "max2", quote(max(data2))),
+    tar_target_raw(name = "mins", quote(c(min1, min2))),
+    tar_target_raw(name = "maxes", quote(c(max1, max2))),
+    tar_target_raw(
+      name = "all",
+      quote(c(browser(), mins, maxes))
+    )
+  )
+)
+# Enter the debugger:
+local_init(pipeline)$run()
+# Run these tests inside the debugger:
+names <- pipeline_get_names(pipeline)
+result <- map_lgl(names, ~is.null(pipeline_get_target(pipeline, .x)$value))
+expect_true(all(result[setdiff(names(result), c("mins", "maxes"))]))
+expect_false(any(result[c("mins", "maxes")]))
+# Exit the debugger.
+
+# Clean up and restart.
+tar_option_reset()
+tar_destroy()
+rstudioapi::restartSession()
+
+# Memory usage tests:
+library(pryr)
 tar_script({
   # Comment out and see memory increase:
   tar_option_set(memory = "transient", garbage_collection = TRUE)
@@ -23,7 +84,6 @@ tar_destroy()
 tar_make()
 
 # The final size of the pipeline object should be small.
-pkgload::load_all()
 tar_destroy()
 tar_option_set(memory = "transient")
 targets <- lapply(
@@ -43,7 +103,6 @@ pryr::object_size(pipeline)
 pryr::object_size(local)
 
 # Pipeline should still be acceptably small even with lots of targets.
-pkgload::load_all()
 tar_destroy()
 pipeline <- pipeline_init(
   list(